+++ port/oc_connectivity.h
@@ -97,9 +97,9 @@ struct oc_message_s
 
 int oc_send_buffer(oc_message_t *message);
 
-int oc_connectivity_init(int device);
+int oc_connectivity_init(size_t device);
 
-void oc_connectivity_shutdown(int device);
+void oc_connectivity_shutdown(size_t device);
 
 void oc_send_discovery_request(oc_message_t *message);
 
@@ -109,7 +109,7 @@ void oc_connectivity_end_session(oc_endpoint_t *endpoint);
 int oc_dns_lookup(const char *domain, oc_string_t *addr, enum transport_flags flags);
 #endif /* OC_DNS_LOOKUP */
 
-oc_endpoint_t *oc_connectivity_get_endpoints(int device);
+oc_endpoint_t *oc_connectivity_get_endpoints(size_t device);
 
 void handle_network_interface_event_callback(oc_interface_event_t event);
